Create AudioDirectorSingleton
